+++ talpid-core/src/dns/android.rs
@@ -10,7 +10,10 @@ pub struct DnsMonitor;
 impl super::DnsMonitorT for DnsMonitor {
     type Error = Error;
 
-    fn new(_cache_dir: impl AsRef<Path>) -> Result<Self, Self::Error> {
+    fn new(
+        _handle: tokio::runtime::Handle,
+        _cache_dir: impl AsRef<Path>,
+    ) -> Result<Self, Self::Error> {
         Ok(DnsMonitor)
     }