+++ talpid-core/src/dns/windows/mod.rs
@@ -50,7 +50,7 @@ pub struct DnsMonitor {}
 impl super::DnsMonitorT for DnsMonitor {
     type Error = Error;
 
-    fn new(cache_dir: impl AsRef<Path>) -> Result<Self, Error> {
+    fn new(_handle: tokio::runtime::Handle, cache_dir: impl AsRef<Path>) -> Result<Self, Error> {
         unsafe { WinDns_Initialize(Some(log_sink), b"WinDns\0".as_ptr()).into_result()? };
 
         let backup_writer = SystemStateWriter::new(